llm_chat_app/requirements.txt
langchain==0.1.17
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.34.0
faiss-cpu==1.8.0
streamlit-extras==0.4.2
pytesseract==0.3.10
pillow==10.3.0
pdf2image==1.17.0
poppler-utils==0.1.0
opencv-python==4.9.0.80
easyocr==1.7.1
openai==1.26.0
tiktoken==0.6.0
langchain-chroma
pinecone
langchain_pinecone
langchain-openai

# brew install poppler or sudo apt-get install poppler-utils

